Fix rounding off inaccuracies

+++ app/controllers/carts_controller.rb
@@ -29,7 +29,7 @@ class CartsController < ApplicationController
     customer_email = params[:stripeEmail]
     Stripe.api_key = "sk_test_L3iQjnqaaxHzHVBzXsgLmR7e"
     Stripe::Charge.create(
-    amount: (@cart.calculate_total*100).to_i,
+    amount: (@cart.calculate_total*100).to_f,
     currency: "usd",
     source: stripe_token
     )
